move plugins/* > plugins.py
