+++ docs/lib/reference/plugins.rst
@@ -1,0 +1,6 @@
+plugins
+=======
+
+
+.. automodule:: starlite.plugins
+    :members:
