--- docs/lib/reference/plugins/base.rst
@@ -1,5 +1,0 @@
-base
-====
-
-.. automodule:: starlite.plugins.base
-    :members:
