--- docs/lib/reference/plugins/index.rst
@@ -1,9 +1,0 @@
-plugins
-=======
-
-.. toctree::
-    :maxdepth: 1
-
-    base
-    piccolo_orm
-    tortoise_orm
