+++ starlite/contrib/sqlalchemy_1/plugin.py
@@ -24,7 +24,7 @@ from starlite.exceptions import (
     ImproperlyConfiguredException,
     MissingDependencyException,
 )
-from starlite.plugins.base import InitPluginProtocol, SerializationPluginProtocol
+from starlite.plugins import InitPluginProtocol, SerializationPluginProtocol
 
 from .types import SQLAlchemyBinaryType
